feat(errors): added error handling config and template
