+++ package.json
@@ -77,12 +77,14 @@
     "handlebars": "^4.0.6",
     "hapi": "^16.0.1",
     "hapi-auth-cookie": "^6.1.1",
+    "hapi-error": "^1.6.0",
     "inert": "^4.0.3",
     "joi": "^10.0.5",
     "materialize-css": "^0.97.8",
     "materialize-loader": "^1.1.1",
     "pg": "^6.1.0",
     "ramda": "^0.22.1",
+    "vision": "^4.1.1",
     "webpack": "^1.13.3"
   },
   "engines": {
