+++ public/login.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Wales Co-op Engagement Tool</title>
+    <link rel="stylesheet" href="./css/styles.css">
+    <link rel="manifest" href="manifest.json">
+    <link rel="icon" type="image/png" href="favicon-32x32.png" sizes="32x32" />
+    <link rel="icon" type="image/png" href="favicon-16x16.png" sizes="16x16" />
+  </head>
+  <body>
+    <nav>
+    <div class="nav-wrapper">
+      <a href="#" class="brand-logo">Test</a>
+      <ul id="nav-mobile" class="right hide-on-med-and-down">
+        <li><a href="">Link</a></li>
+      </ul>
+    </div>
+  </nav> 
+    <script src="bundle.js"></script>
+  </body>
+</html>
