+++ src/server/index.js
@@ -1,8 +1,7 @@
 import hapi from 'hapi';
 import Path from 'path';
-import plugins from './plugins';
+import plugins, * as plug from './plugins';
 import routes from './routes';
-import { baseConfig } from './auth';
 import logger from './logger';
 
 const server = new hapi.Server({
@@ -21,7 +20,8 @@ server.connection({
 
 server.register(plugins, (err) => {
   if (err) throw err;
-  server.auth.strategy('base', 'cookie', 'required', baseConfig);
+  server.auth.strategy('base', 'cookie', 'required', plug.authConfig);
+  server.views(plug.viewsConfig);
   server.route(routes);
   server.start(() =>
     // eslint-disable-next-line no-console
