+++ src/views/errors.hbs
@@ -1,0 +1,12 @@
+<div class="row">
+  <div class="col s12">
+    <p class="flat-text-logo center white-text caption-uppercase">{{errorTitle}}</p>
+  </div>
+  <div class="row center">
+    <h1 class="text-long-shadow col s12">{{statusCode}}</h1>
+  </div>
+  <div class="row center">
+    <p class="center white-text col s12">{{errorMessage}}</p>
+  </div>
+</div>
+
