+++ src/views/layout/default.hbs
@@ -1,0 +1,1 @@
+// layout view goes here
